requirements.txt
streamlit>=1.41.0
python-dotenv>=1.0.0
pandas>=2.2.0
wirl-lang>=0.1.0
wirl-pregel-runner>=0.1.0
langchain-ollama>=0.2.1
langchain-openai>=0.2.14
numpy>=1.24.0
langgraph>=0.2.60
pytest>=8.0.0
